fix(financial): harden TON SDK integration — 11 audit fixes with TDD
CRITICAL fixes:
- Add CAS version check to updateConfirmed() preventing double-spend
- Filter null txHash/inMsg from TON Center API responses
- Eliminate mnemonic exposure in exception stack traces

HIGH fixes:
- Change getSeqno() return type from int to long (overflow prevention)
- Add subwalletId overflow guard in EscrowService
- Validate positive amount in DepositWatcher matching
- Select most recent TX by lt instead of findFirst()

MEDIUM fixes:
- Add retry_count column and failure escalation (DB migration 019)
- Add maxRetries property to TonProperties.Deposit
- Add TON_DEPOSIT_FAILED metric

Tests:
- 20+ new unit tests covering all fixes
- Negative tests for TonCenterBlockchainAdapter (ok=false, empty result)
- Idempotency tests for EscrowService (confirm/release/refund)

+++ advert-market-financial-api/src/main/java/com/advertmarket/financial/api/port/TonBlockchainPort.java
@@ -25,7 +25,7 @@ public interface TonBlockchainPort {
     long getAddressBalance(@NonNull String address);
 
     /** Returns the current seqno of a wallet contract. */
-    int getSeqno(@NonNull String address);
+    long getSeqno(@NonNull String address);
 
     /** Estimates transaction fee in nanoTON. */
     long estimateFee(@NonNull String address, @NonNull String base64Body);
